Move InterpolationWeights for use in Spectral w/o circular dep
