+++ tests/Unit/NumericalAlgorithms/Interpolation/CMakeLists.txt
@@ -9,7 +9,6 @@ set(LIBRARY_SOURCES
   Test_BarycentricRational.cpp
   Test_CardinalInterpolator.cpp
   Test_CubicSpline.cpp
-  Test_InterpolationWeights.cpp
   Test_IrregularInterpolant.cpp
   Test_LagrangePolynomial.cpp
   Test_MultiLinearSpanInterpolation.cpp
