+++ tests/Unit/NumericalAlgorithms/Spectral/CMakeLists.txt
@@ -24,6 +24,7 @@ set(LIBRARY_SOURCES
   Test_IndefiniteIntegral.cpp
   Test_IntegrationMatrix.cpp
   Test_InterpolationMatrix.cpp
+  Test_InterpolationWeights.cpp
   Test_Legendre.cpp
   Test_LegendreGauss.cpp
   Test_LegendreGaussLobatto.cpp
